Improve windows support of relative_to
The documentation for windows paths has been improved
showing examples where it is not possible to compute
a relative path and how to get an absolute path in that
situation.

Since the computation of relative paths is system dependent,
relative_to now accepts a "native" argument to use either
the build (native: true) or the host (native: false, default)
system as reference.

Tests for windows paths are now available and error messages
have been clarified as well.